pyproject.toml
[project]
name = "rag-search-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "nltk==3.9.1",
    "python-dotenv>=1.2.2",
]
[project.scripts]
ksearch = "cli.keyword_search_cli:main"

[tool.setuptools.packages.find]
include = ["cli*"]

[dependency-groups]
dev = [
    "ipython>=9.10.1",
    "ruff>=0.15.8",
]
